Fix circular import for `PageInterface`
