+++ tests/testapp/mutations.py
@@ -3,7 +3,7 @@ import graphene
 from wagtail.models import Page
 
 from grapple.registry import registry
-from grapple.types.pages import PageInterface
+from grapple.types.interfaces import PageInterface
 from grapple.types.rich_text import RichText
 from testapp.models import Advert, AuthorPage
